requirements.txt
gradio==4.5.0
mmpose==1.2.0
moviepy==1.0.3
numpy==1.24.4
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==10.1.0
